removed uneeded data loader
